add first version to the operations percentage

+++ src/main/java/com/magdy/challenge/tenpo/infrastructure/delivery/SumRestController.java
@@ -16,7 +16,7 @@ public class SumRestController {
     }
 
     @GetMapping
-    public Integer sum(int value1 , int value2){
+    public float sum(int value1 , int value2){
         return sumEndpoints.sum(value1,value2);
     }
 }
